Finish up testing
