+++ crates/holochain/src/core/workflow/validation_receipt_workflow/tests.rs
@@ -1,238 +1,316 @@
-use crate::core::queue_consumer::WorkComplete;
-use crate::core::workflow::validation_receipt_workflow::validation_receipt_workflow;
-use crate::prelude::AgentPubKeyFixturator;
-use crate::prelude::CreateFixturator;
-use crate::prelude::DhtOpHashed;
-use crate::prelude::SignatureFixturator;
-use fixt::fixt;
-use futures::future::BoxFuture;
-use futures::FutureExt;
-use hdk::prelude::Action;
-use holo_hash::fixt::DnaHashFixturator;
-use holo_hash::HasHash;
-use holo_hash::HoloHashOf;
-use holo_hash::{AgentPubKey, DhtOpHash};
-use holochain_p2p::MockHolochainP2pDnaT;
-use holochain_sqlite::error::DatabaseResult;
-use holochain_sqlite::prelude::{DbKindDht, DbWrite};
-use holochain_state::prelude::set_require_receipt;
-use holochain_state::prelude::set_validation_status;
-use holochain_state::prelude::set_when_integrated;
-use holochain_state::prelude::StateMutationResult;
-use holochain_types::dht_op::DhtOp;
-use holochain_types::prelude::Timestamp;
-use holochain_zome_types::cell::CellId;
-use holochain_zome_types::prelude::ValidationStatus;
-use holochain_zome_types::{Block, BlockTarget, CellBlockReason};
-use parking_lot::RwLock;
+use crate::conductor::api::error::ConductorApiResult;
+use crate::core::ribosome::guest_callback::validate::ValidateResult;
+use crate::prelude::InlineZomeSet;
+use crate::sweettest::*;
+use crate::test_utils::consistency_10s;
+use crate::test_utils::inline_zomes::simple_create_read_zome;
+use hdk::prelude::*;
+use holo_hash::DhtOpHash;
+use holochain_keystore::AgentPubKeyExt;
+use holochain_state::prelude::*;
+use holochain_types::prelude::block::BlockTargetId;
+use holochain_types::prelude::*;
 use rusqlite::Transaction;
-use std::sync::Arc;
 
 #[tokio::test(flavor = "multi_thread")]
-async fn no_running_cells() {
-    holochain_trace::test_run().ok();
+#[ignore = "flaky"]
+async fn test_validation_receipt() {
+    let _g = holochain_trace::test_run().ok();
+    const NUM_CONDUCTORS: usize = 3;
 
-    let test_db = holochain_state::test_utils::test_dht_db();
-    let vault = test_db.to_db();
-    let keystore = holochain_state::test_utils::test_keystore();
+    let mut conductors = SweetConductorBatch::from_standard_config(NUM_CONDUCTORS).await;
 
-    let mut dna = MockHolochainP2pDnaT::new();
-    dna.expect_send_validation_receipt().never(); // Verify no receipts sent
+    let (dna_file, _, _) =
+        SweetDnaFile::unique_from_inline_zomes(("simple", simple_create_read_zome())).await;
 
-    let work_complete = validation_receipt_workflow(
-        Arc::new(fixt!(DnaHash)),
-        vault,
-        dna,
-        keystore,
-        vec![].into_iter().collect(), // No running cells
-        |_block| unreachable!("This test should not send a block"),
-    )
-    .await
-    .unwrap();
+    let apps = conductors.setup_app("app", &[dna_file]).await.unwrap();
+    conductors.exchange_peer_info().await;
 
-    assert_eq!(WorkComplete::Complete, work_complete);
-}
+    let ((alice,), (bobbo,), (carol,)) = apps.into_tuples();
 
-#[tokio::test(flavor = "multi_thread")]
-async fn do_not_block_or_send_to_self() {
-    holochain_trace::test_run().ok();
+    // Call the "create" zome fn on Alice's app
+    let hash: ActionHash = conductors[0]
+        .call(&alice.zome("simple"), "create", ())
+        .await;
 
-    let test_db = holochain_state::test_utils::test_dht_db();
-    let vault = test_db.to_db();
-    let keystore = holochain_state::test_utils::test_keystore();
+    consistency_10s([&alice, &bobbo, &carol]).await;
 
-    let dna_hash = fixt!(DnaHash);
-    let author = fixt!(AgentPubKey);
+    // Get op hashes
+    let vault = alice.dht_db();
+    let record = vault
+        .read_async(move |txn| -> StateQueryResult<Record> {
+            Ok(Txn::from(&txn).get_record(&hash.clone().into())?.unwrap())
+        })
+        .await
+        .unwrap();
+    let ops = produce_ops_from_record(&record)
+        .unwrap()
+        .into_iter()
+        .map(|op| DhtOpHash::with_data_sync(&op))
+        .collect::<Vec<_>>();
 
-    // Create a valid op that would require a validation receipt except that it's created by us
-    create_modified_op(vault.clone(), Some(author.clone()), |txn, op_hash| {
-        set_require_receipt(txn, &op_hash, true)?;
-        set_when_integrated(txn, &op_hash, Timestamp::now())?;
-        set_validation_status(txn, &op_hash, ValidationStatus::Valid)?;
-        Ok(())
-    })
-    .await
-    .unwrap();
-
-    // Create a rejected op which would usually cause a block but it's created by us
-    create_modified_op(vault.clone(), Some(author.clone()), |txn, op_hash| {
-        set_require_receipt(txn, &op_hash, true)?;
-        set_when_integrated(txn, &op_hash, Timestamp::now())?;
-        set_validation_status(txn, &op_hash, ValidationStatus::Rejected)?;
-        Ok(())
-    })
-    .await
-    .unwrap();
-
-    let mut dna = MockHolochainP2pDnaT::new();
-
-    dna.expect_send_validation_receipt().never(); // Verify no receipts sent
-
-    let validator = CellId::new(dna_hash.clone(), author);
-
-    let work_complete = validation_receipt_workflow(
-        Arc::new(dna_hash),
-        vault,
-        dna,
-        keystore,
-        vec![validator].into_iter().collect(), // No running cells
-        |_block| unreachable!("This test should not send a block"), // Verify no blocks sent
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(WorkComplete::Complete, work_complete);
-}
-
-#[tokio::test(flavor = "multi_thread")]
-async fn block_invalid_op_author() {
-    holochain_trace::test_run().ok();
-
-    let test_db = holochain_state::test_utils::test_dht_db();
-    let vault = test_db.to_db();
-    let keystore = holochain_state::test_utils::test_keystore();
-
-    // Any op created by somebody else, which has been rejected by validation.
-    let (author, op_hash) = create_modified_op(vault.clone(), None, |txn, op_hash| {
-        set_require_receipt(txn, &op_hash, true)?;
-        set_when_integrated(txn, &op_hash, Timestamp::now())?;
-        set_validation_status(txn, &op_hash, ValidationStatus::Rejected)?;
-        Ok(())
-    })
-    .await
-    .unwrap();
-
-    // We'll still send a validation receipt, but we should also block them
-    let mut dna = MockHolochainP2pDnaT::new();
-    dna.expect_send_validation_receipt()
-        .return_once(|_, _| Ok(()));
-
-    let dna_hash = fixt!(DnaHash);
-    let validator = CellId::new(
-        dna_hash.clone(),
-        keystore.new_sign_keypair_random().await.unwrap(),
+    // Wait for receipts to be sent
+    crate::assert_eq_retry_10s!(
+        {
+            let mut counts = Vec::new();
+            for hash in &ops {
+                let count = vault
+                    .read_async({
+                        let query_hash = hash.clone();
+                        move |r| -> StateQueryResult<usize> {
+                            Ok(list_receipts(&r, &query_hash)?.len())
+                        }
+                    })
+                    .await
+                    .unwrap();
+                counts.push(count);
+            }
+            counts
+        },
+        vec![2, 2, 2],
     );
 
-    let blocks = Arc::new(RwLock::new(Vec::<Block>::new()));
-
-    let work_complete = validation_receipt_workflow(
-        Arc::new(dna_hash),
-        vault,
-        dna,
-        keystore,
-        vec![validator].into_iter().collect(),
-        {
-            let blocks = blocks.clone();
-            move |block| -> BoxFuture<DatabaseResult<()>> {
-                blocks.write().push(block);
-                async move { Ok(()) }.boxed().into()
-            }
-        },
-    )
-    .await
-    .unwrap();
-
-    assert_eq!(WorkComplete::Complete, work_complete);
-
-    let read_blocks = blocks.read();
-    assert_eq!(1, read_blocks.len());
-    match read_blocks.first().unwrap().target() {
-        BlockTarget::Cell(cell_id, reason) => {
-            assert_eq!(CellBlockReason::InvalidOp(op_hash), *reason);
-            assert_eq!(author, *cell_id.agent_pubkey());
+    // Check alice has receipts from both bobbo and carol
+    for hash in &ops {
+        let receipts: Vec<_> = vault
+            .read_async({
+                let query_hash = hash.clone();
+                move |r| list_receipts(&r, &query_hash)
+            })
+            .await
+            .unwrap();
+        assert_eq!(receipts.len(), 2);
+        for receipt in receipts {
+            let SignedValidationReceipt {
+                receipt,
+                validators_signatures: sigs,
+            } = receipt;
+            let validator = receipt.validators[0].clone();
+            assert!(validator == *bobbo.agent_pubkey() || validator == *carol.agent_pubkey());
+            assert!(validator.verify_signature(&sigs[0], receipt).await.unwrap());
         }
-        _ => unreachable!("Only expect a cell block"),
+    }
+
+    // Check alice has 2 receipts in their authored dht ops table.
+    crate::assert_eq_retry_1m!(
+        {
+            vault
+                .read_async(move |txn: Transaction| -> DatabaseResult<Vec<u32>> {
+                    let mut stmt = txn
+                        .prepare("SELECT COUNT(hash) FROM ValidationReceipt GROUP BY op_hash")
+                        .unwrap();
+                    Ok(stmt
+                        .query_map([], |row| row.get::<_, Option<u32>>(0))
+                        .unwrap()
+                        .map(Result::unwrap)
+                        .filter_map(|i| i)
+                        .collect::<Vec<u32>>())
+                })
+                .await
+                .unwrap()
+        },
+        vec![2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2]
+    );
+}
+
+macro_rules! wait_until {
+    ($expression:expr; $interval_ms:literal; $timeout_ms:literal; $wait_msg:literal; $timeout_msg:literal;) => {
+        let timeout = (Timestamp::now() + std::time::Duration::from_millis($timeout_ms)).unwrap();
+        let interval_duration = std::time::Duration::from_millis($interval_ms);
+        tokio::time::sleep(interval_duration).await;
+        while !$expression {
+            if Timestamp::now() > timeout {
+                panic!($timeout_msg);
+            }
+            dbg!($wait_msg);
+            tokio::time::sleep(interval_duration).await;
+        }
+    };
+}
+
+#[tokio::test(flavor = "multi_thread")]
+#[cfg_attr(target_os = "macos", ignore = "flaky")]
+async fn test_block_invalid_receipt() {
+    holochain_trace::test_run().ok();
+    let unit_entry_def = EntryDef::default_from_id("unit");
+    let integrity_name = "integrity";
+    let coordinator_name = "coordinator";
+    let integrity_uuid = "a";
+    let create_coordinator_uuid = "b";
+    let check_coordinator_uuid = "c";
+    let network_seed = "network_seed";
+    let create_function_name = "create";
+    let app_prefix = "app-";
+
+    let zomes_that_create = InlineZomeSet::new_single(
+        integrity_name,
+        coordinator_name,
+        integrity_uuid,
+        create_coordinator_uuid,
+        vec![unit_entry_def.clone()],
+        0,
+    )
+    .function(coordinator_name, create_function_name, move |api, ()| {
+        let entry = Entry::app(().try_into().unwrap()).unwrap();
+        let hash = api.create(CreateInput::new(
+            InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
+            EntryVisibility::Public,
+            entry,
+            ChainTopOrdering::default(),
+        ))?;
+        Ok(hash)
+    });
+    // .function(
+    //     coordinator_name,
+    //     get_function_name,
+    //     move |api, hash: AnyDhtHash| {
+    //         let records = api.get(vec![GetInput::new(hash, Default::default())])?;
+    //         Ok(records[0])
+    //     },
+    // );
+
+    let zomes_that_check = InlineZomeSet::new_single(
+        integrity_name,
+        coordinator_name,
+        integrity_uuid,
+        check_coordinator_uuid,
+        vec![unit_entry_def.clone()],
+        0,
+    )
+    .function(integrity_name, "validate", |_api, op: Op| match op {
+        Op::StoreEntry(StoreEntry { action, .. })
+            if action.hashed.content.app_entry_def().is_some() =>
+        {
+            dbg!("entry defs ARE bad!");
+            Ok(ValidateResult::Invalid("Entry defs are bad".into()))
+        }
+        _ => Ok(ValidateResult::Valid),
+    });
+
+    let config = SweetConductorConfig::rendezvous();
+    let conductors = SweetConductorBatch::from_config_rendezvous(2, config).await;
+
+    let mut conductors = conductors.into_inner().into_iter();
+
+    let mut alice_conductor = conductors.next().unwrap();
+    let mut bob_conductor = conductors.next().unwrap();
+    let (alice_pubkey, bob_pubkey) = SweetAgents::alice_and_bob();
+
+    let (dna_that_creates, _, _) =
+        SweetDnaFile::from_inline_zomes(network_seed.into(), zomes_that_create).await;
+
+    let (dna_that_checks, _, _) =
+        SweetDnaFile::from_inline_zomes(network_seed.into(), zomes_that_check).await;
+
+    let alice_apps = alice_conductor
+        .setup_app_for_agents(app_prefix, &[alice_pubkey.clone()], &[dna_that_creates])
+        .await
+        .unwrap();
+
+    let ((alice_cell,),) = alice_apps.into_tuples();
+
+    let bob_apps = bob_conductor
+        .setup_app_for_agents(app_prefix, &[bob_pubkey.clone()], &[dna_that_checks])
+        .await
+        .unwrap();
+
+    let ((bob_cell,),) = bob_apps.into_tuples();
+
+    let _action_hash: ActionHash = alice_conductor
+        .call(&alice_cell.zome(coordinator_name), create_function_name, ())
+        .await;
+
+    consistency_10s([&alice_cell, &bob_cell]).await;
+
+    let alice_block_target = BlockTargetId::Cell(alice_cell.cell_id().to_owned());
+    let bob_block_target = BlockTargetId::Cell(bob_cell.cell_id().to_owned());
+
+    for now in vec![Timestamp::now(), Timestamp::MIN, Timestamp::MAX] {
+        assert!(!alice_conductor
+            .spaces
+            .is_blocked(alice_block_target.clone(), now)
+            .await
+            .unwrap());
+        assert!(!alice_conductor
+            .spaces
+            .is_blocked(bob_block_target.clone(), now)
+            .await
+            .unwrap());
+        assert!(!bob_conductor
+            .spaces
+            .is_blocked(bob_block_target.clone(), now)
+            .await
+            .unwrap());
+
+        // It can take a little more than consistency to have the receipts
+        // processed.
+        wait_until!(
+            bob_conductor.spaces.is_blocked(alice_block_target.clone(), now).await.unwrap();
+            1000;
+            10000;
+            "waiting for block due to warrant";
+            "warrant block never happened";
+        );
     }
 }
 
 #[tokio::test(flavor = "multi_thread")]
-async fn stops_if_receipt_cannot_be_signed() {
+async fn test_not_block_self_receipt() {
     holochain_trace::test_run().ok();
 
-    let test_db = holochain_state::test_utils::test_dht_db();
-    let vault = test_db.to_db();
-    let keystore = holochain_state::test_utils::test_keystore();
-
-    // Any op created by somebody else, which is valid
-    let (author, op_hash) = create_modified_op(vault.clone(), None, |txn, op_hash| {
-        set_require_receipt(txn, &op_hash, true)?;
-        set_when_integrated(txn, &op_hash, Timestamp::now())?;
-        set_validation_status(txn, &op_hash, ValidationStatus::Valid)?;
-        Ok(())
-    })
-    .await
-    .unwrap();
-
-    let mut dna = MockHolochainP2pDnaT::new();
-    dna.expect_send_validation_receipt().never();
-
-    let dna_hash = fixt!(DnaHash);
-
-    let invalid_validator = CellId::new(
-        dna_hash.clone(),
-        fixt!(AgentPubKey), // Not valid because it won't be found in Lair
-    );
-
-    let work_complete = validation_receipt_workflow(
-        Arc::new(dna_hash),
-        vault,
-        dna,
-        keystore,
-        vec![invalid_validator].into_iter().collect(), // No running cells
-        |_block| unreachable!("Should not try to block"),
+    let unit_entry_def = EntryDef::default_from_id("unit");
+    let zomes = InlineZomeSet::new_single(
+        "integrity",
+        "coordinator",
+        "a",
+        "b",
+        vec![unit_entry_def.clone()],
+        0,
     )
-    .await
-    .unwrap();
+    .function("coordinator", "create", move |api, ()| {
+        let entry = Entry::app(().try_into().unwrap()).unwrap();
+        let hash = api.create(CreateInput::new(
+            InlineZomeSet::get_entry_location(&api, EntryDefIndex(0)),
+            EntryVisibility::Public,
+            entry,
+            ChainTopOrdering::default(),
+        ))?;
+        Ok(hash)
+    })
+    .function("integrity", "validate", |_api, op: Op| match op {
+        Op::StoreEntry(StoreEntry { action, .. })
+            if action.hashed.content.app_entry_def().is_some() =>
+        {
+            Ok(ValidateResult::Invalid("Entry defs are bad".into()))
+        }
+        _ => Ok(ValidateResult::Valid),
+    });
 
-    assert_eq!(WorkComplete::Incomplete, work_complete);
-}
+    let mut conductor = SweetConductor::from_standard_config().await;
+    let alice_pubkey = SweetAgents::alice();
+    let (dna, _, _) = SweetDnaFile::from_inline_zomes("network_seed".into(), zomes).await;
 
-async fn create_modified_op(
-    vault: DbWrite<DbKindDht>,
-    author: Option<AgentPubKey>,
-    modifier: fn(txn: &mut Transaction, op_hash: HoloHashOf<DhtOp>) -> StateMutationResult<()>,
-) -> StateMutationResult<(AgentPubKey, DhtOpHash)> {
-    // The actual op does not matter, just some of the status fields
-    let mut create_action = fixt!(Create);
-    let author = author.unwrap_or_else(|| fixt!(AgentPubKey));
-    create_action.author = author.clone();
-    let action = Action::Create(create_action);
-
-    let op = DhtOpHashed::from_content_sync(DhtOp::RegisterAgentActivity(fixt!(Signature), action));
-
-    let test_op_hash = op.as_hash().clone();
-    vault
-        .write_async({
-            let test_op_hash = test_op_hash.clone();
-            move |txn| -> StateMutationResult<()> {
-                holochain_state::mutations::insert_op(txn, &op)?;
-                modifier(txn, test_op_hash)?;
-
-                Ok(())
-            }
-        })
+    let apps = conductor
+        .setup_app_for_agents("app-", &[alice_pubkey.clone()], &[dna])
         .await
         .unwrap();
 
-    Ok((author, test_op_hash))
+    let ((alice_cell,),) = apps.into_tuples();
+
+    let maybe_action_hash: ConductorApiResult<ActionHash> = conductor
+        .call_fallible(&alice_cell.zome("coordinator"), "create", ())
+        .await;
+    assert!(maybe_action_hash.is_err());
+
+    let alice_block_target = BlockTargetId::Cell(alice_cell.cell_id().to_owned());
+
+    // Give some time to ensure alice clears workflows.
+    tokio::time::sleep(std::time::Duration::from_millis(1000)).await;
+
+    // Alice should not block herself for finding her own invalid entry.
+    assert!(!conductor
+        .spaces
+        .is_blocked(alice_block_target, Timestamp::now())
+        .await
+        .unwrap());
 }
